requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
playsound==1.3.0
pyrepl==0.11.3.post1
qrcode==8.2
setuptools==80.9.0
style==1.1.0
update==0.0.1
Werkzeug==3.1.3
wheel==0.45.1
